requirements.txt
django>=5.0,<6.0
aiogram>=3.0,<4.0
celery[redis]>=5.3.0
python-dotenv>=1.0.0
django-celery-beat>=2.5.0
django-grappelli>=3.1.0
pytz>=2024.0
timezonefinder>=6.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
